acs: change noise amplifier coefficient to 2
We change the noise amplifier coffecient  to 2 so we  now get:

                    (busy time - tx time)
    int_factor = ------------------------- *  2^(noise + min_noise)
                   (active time - tx time)

The coefficient of of 2 reflects the way power in "far-field" radiation
decreases as the square of distance from the antenna [1]. Thanks to
Felix Bitterli for the hint.

This reduction also means smaller results. If we want even smaller
values we can use log2() for the results but not sure what is best.

[1] http://en.wikipedia.org/wiki/Near_and_far_field

+++ acs.h
@@ -32,8 +32,8 @@ struct freq_item {
 	bool enabled;
 	__s8 max_noise;
 	__s8 min_noise;
-	/* An alternative is to use__float128 for low noise environments */
-	long long unsigned int interference_factor;
+	/* An alternative is to use __float128 for low noise environments */
+	long double interference_factor;
 	struct list_head list_member;
 	struct list_head survey_list;
 };
